Add store command
