+++ cg/cli/workflow/nf_analysis.py
@@ -143,3 +143,35 @@ def store_housekeeper(context: CGConfig, case_id: str, dry_run: bool) -> None:
     except HousekeeperStoreError as error:
         LOG.error(f"Could not store bundle in Housekeeper and StatusDB: {error}!")
         raise click.Abort()
+
+
+@click.command("store")
+@ARGUMENT_CASE_ID
+@OPTION_DRY
+@click.pass_context
+def store(context: click.Context, case_id: str, dry_run: bool) -> None:
+    """Generate deliverables files for a case and store in Housekeeper if they
+    pass QC metrics checks."""
+    analysis_api: NfAnalysisAPI = context.obj.meta_apis[MetaApis.ANALYSIS_API]
+
+    is_latest_analysis_qc: bool = analysis_api.trailblazer_api.is_latest_analysis_qc(
+        case_id=case_id
+    )
+    if not is_latest_analysis_qc and not analysis_api.trailblazer_api.is_latest_analysis_completed(
+        case_id=case_id
+    ):
+        LOG.error(
+            "Case not stored. Trailblazer status must be either QC or COMPLETE to be able to store"
+        )
+        return
+
+    # Avoid storing a case without QC checks previously performed
+    if (
+        is_latest_analysis_qc
+        or not analysis_api.get_metrics_deliverables_path(case_id=case_id).exists()
+    ):
+        LOG.info(f"Generating metrics file and performing QC checks for {case_id}")
+        context.invoke(metrics_deliver, case_id=case_id, dry_run=dry_run)
+    LOG.info(f"Storing analysis for {case_id}")
+    context.invoke(report_deliver, case_id=case_id, dry_run=dry_run)
+    context.invoke(store_housekeeper, case_id=case_id, dry_run=dry_run)
