+++ cg/cli/workflow/rnafusion/base.py
@@ -6,7 +6,6 @@ from pathlib import Path
 import click
 from pydantic.v1 import ValidationError
 
-from cg.apps.housekeeper.hk import HousekeeperAPI
 from cg.cli.workflow.commands import ARGUMENT_CASE_ID, resolve_compression
 from cg.cli.workflow.nf_analysis import (
     OPTION_COMPUTE_ENV,
@@ -22,6 +21,7 @@ from cg.cli.workflow.nf_analysis import (
     metrics_deliver,
     report_deliver,
     store_housekeeper,
+    store,
 )
 from cg.cli.workflow.rnafusion.options import (
     OPTION_REFERENCES,
@@ -225,38 +225,7 @@ def start_available(context: click.Context, dry_run: bool = False) -> None:
 rnafusion.add_command(metrics_deliver)
 rnafusion.add_command(report_deliver)
 rnafusion.add_command(store_housekeeper)
-
-
-@rnafusion.command("store")
-@ARGUMENT_CASE_ID
-@DRY_RUN
-@click.pass_context
-def store(context: click.Context, case_id: str, dry_run: bool) -> None:
-    """Generate deliverables files for a case and store in Housekeeper if they
-    pass QC metrics checks."""
-    analysis_api: RnafusionAnalysisAPI = context.obj.meta_apis[MetaApis.ANALYSIS_API]
-
-    is_latest_analysis_qc: bool = analysis_api.trailblazer_api.is_latest_analysis_qc(
-        case_id=case_id
-    )
-    if not is_latest_analysis_qc and not analysis_api.trailblazer_api.is_latest_analysis_completed(
-        case_id=case_id
-    ):
-        LOG.error(
-            "Case not stored. Trailblazer status must be either QC or COMPLETE to be able to store"
-        )
-        return
-
-    # Avoid storing a case without QC checks previously performed
-    if (
-        is_latest_analysis_qc
-        or not analysis_api.get_metrics_deliverables_path(case_id=case_id).exists()
-    ):
-        LOG.info(f"Generating metrics file and performing QC checks for {case_id}")
-        context.invoke(metrics_deliver, case_id=case_id, dry_run=dry_run)
-    LOG.info(f"Storing analysis for {case_id}")
-    context.invoke(report_deliver, case_id=case_id, dry_run=dry_run)
-    context.invoke(store_housekeeper, case_id=case_id, dry_run=dry_run)
+rnafusion.add_command(store)
 
 
 @rnafusion.command("store-available")
