+++ tests/cli/workflow/rnafusion/test_cli_rnafusion_compound_commands.py
@@ -64,95 +64,6 @@ def test_start(
     assert "-resume" not in caplog.text
 
 
-def test_store_success(
-    cli_runner: CliRunner,
-    rnafusion_context: CGConfig,
-    real_housekeeper_api: HousekeeperAPI,
-    rnafusion_mock_deliverable_dir,
-    rnafusion_mock_analysis_finish,
-    caplog: LogCaptureFixture,
-    rnafusion_hermes_deliverables,
-    mocker,
-    rnafusion_case_id: str,
-    deliverables_template_content: list[dict],
-):
-    """Test to ensure all parts of store command are run successfully given ideal conditions."""
-    caplog.set_level(logging.INFO)
-
-    # GIVEN a case for which we mocked files created after a successful run
-
-    # GIVEN a mocked deliverables template
-    mocker.patch.object(
-        RnafusionAnalysisAPI,
-        "get_deliverables_template_content",
-        return_value=deliverables_template_content,
-    )
-
-    # Set Housekeeper to an empty real Housekeeper store
-    rnafusion_context.housekeeper_api_: HousekeeperAPI = real_housekeeper_api
-    rnafusion_context.meta_apis["analysis_api"].housekeeper_api = real_housekeeper_api
-
-    # Make sure the bundle was not present in hk
-    assert not rnafusion_context.housekeeper_api.bundle(rnafusion_case_id)
-
-    # Make sure analysis not already stored in status_db
-    assert not rnafusion_context.status_db.get_case_by_internal_id(
-        internal_id=rnafusion_case_id
-    ).analyses
-
-    # GIVEN that HermesAPI returns a deliverables output
-    mocker.patch.object(HermesApi, "convert_deliverables")
-    HermesApi.convert_deliverables.return_value = CGDeliverables(**rnafusion_hermes_deliverables)
-
-    # WHEN running command
-    result = cli_runner.invoke(store, [rnafusion_case_id], obj=rnafusion_context)
-
-    # THEN bundle should be successfully added to Housekeeper and StatusDB
-    assert result.exit_code == EXIT_SUCCESS
-    assert "Analysis successfully stored in Housekeeper" in caplog.text
-    assert "Analysis successfully stored in StatusDB" in caplog.text
-    assert rnafusion_context.status_db.get_case_by_internal_id(
-        internal_id=rnafusion_case_id
-    ).analyses
-    assert rnafusion_context.housekeeper_api.bundle(rnafusion_case_id)
-
-
-def test_store_fail(
-    cli_runner: CliRunner,
-    rnafusion_case_id: str,
-    rnafusion_context: CGConfig,
-    real_housekeeper_api: HousekeeperAPI,
-    deliverables_template_content: list[dict],
-    caplog: LogCaptureFixture,
-    mocker,
-):
-    """Test store command fails when a case did not finish."""
-    caplog.set_level(logging.INFO)
-
-    # GIVEN CASE ID where analysis finish is not mocked
-    case_id_fail: str = rnafusion_case_id
-
-    # GIVEN a mocked deliverables template
-    mocker.patch.object(
-        RnafusionAnalysisAPI,
-        "get_deliverables_template_content",
-        return_value=deliverables_template_content,
-    )
-
-    # Set Housekeeper to an empty real Housekeeper store
-    rnafusion_context.housekeeper_api_: HousekeeperAPI = real_housekeeper_api
-    rnafusion_context.meta_apis["analysis_api"].housekeeper_api = real_housekeeper_api
-
-    # Make sure the bundle was not present in hk
-    assert not rnafusion_context.housekeeper_api.bundle(case_id_fail)
-
-    # WHEN running command
-    result_fail = cli_runner.invoke(store, [case_id_fail], obj=rnafusion_context)
-
-    # THEN bundle exist status should be
-    assert result_fail.exit_code != EXIT_SUCCESS
-
-
 def test_start_available(
     cli_runner: CliRunner,
     rnafusion_context: CGConfig,
