+++ tests/cli/workflow/test_cli_workflow.py
@@ -22,3 +22,4 @@ def test_no_options(cli_runner: CliRunner, base_context: CGConfig):
     assert "mip-dna" in result.output
     assert "mip-rna" in result.output
     assert "rnafusion" in result.output
+    assert "taxprofiler" in result.output
